Set dtypes for integer, Windows Long Long problem again

+++ avaframe/com1DFA/damCom1DFA.pyx
@@ -558,7 +558,8 @@ def initializeWallLines(cfg, dem, wallLineDict, savePath=''):
     #   fileName = shpConv.writeLine2SHPfile(wallLineDict, 'dam foot line', savePath, header=dem['originalHeader'])
   else:
     # crete a dummy dict (needed so that cython runs)
-    wallLineDict = {'dam': 0, 'cellsCrossed': np.zeros((dem['header']['ncols']*dem['header']['nrows'])).astype(int)}
+    wallLineDict = {'dam': 0, 'cellsCrossed': np.zeros((dem['header']['ncols']*dem['header']['nrows'])).astype(
+        np.int32)}
     for key in ['x', 'y', 'z', 'xCrown', 'yCrown', 'zCrown', 'xTangent', 'yTangent', 'zTangent']:
       wallLineDict[key] = np.ones((1))*1.0
     for key in ['nPoints', 'height', 'slope', 'restitutionCoefficient', 'nIterDam']:
